+++ lib/recode/task/pipe_fun_one.ex
@@ -11,6 +11,10 @@ defmodule Recode.Task.PipeFunOne do
   This task rewrites the code when `mix recode` runs with `autocorrect: true`.
   """
 
+  @shortdoc "Add parentheses to one-arity functions."
+
+  @category :readability
+
   use Recode.Task, correct: true, check: true
 
   alias Recode.Issue