+++ lib/mix/tasks/recode.ex
@@ -112,7 +112,7 @@ defmodule Mix.Tasks.Recode do
         config
 
       {:error, :out_of_date} ->
-        Mix.raise("The config is out of date. Run `mix recode.gen.config` to update.")
+        Mix.raise("The config is out of date. Run `mix recode.update.config` to update.")
 
       {:error, :no_tasks} ->
         Mix.raise("No `:tasks` key found in configuration.")