+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 + Add `Recode.Task.FilterCount`.
 + Refactor `RecodeCase`
 + Use switch `--task` multiple times.
++ Add mix task `recode.help`.
 
 ## 0.5.2 - 2023/07/17